requirements.txt
aiogram==2.22.2
nest_asyncio==1.5.6
python-decouple==3.6
